Split field offset to its subparts
According to GCC docs:
The position of the field within the parent record is specified by a combination
of three attributes. DECL_FIELD_OFFSET is the position, counting in bytes, of
the DECL_OFFSET_ALIGN-bit sized word containing the bit of the field closest to
the beginning of the structure. DECL_FIELD_BIT_OFFSET is the bit offset of the
first bit of the field within this word